Correction of Missing FilenamePrefix
Because the Gui dont provide a custom Filename the filenamePrefixStr is always empty. Because of this the cmake creation always use the defined Standard Prefix "myPlugin". This results in bad Filenames after creation. As long the GUI not provide a definition of a custom filename this change copy the Classname to the FilenamePrefix.